Move Asset class to PSR-4 location
This sets the asset class within the Models folder for PSR-4 loading.
