+++ src/lib/models/stack.php
@@ -1,11 +1,10 @@
 <?php
 namespace Contentstack\Stack;
 
+use Contentstack\Models\Assets;
 use Contentstack\Support\Utility;
-use Contentstack\Stack\Assets\Assets;
 use Contentstack\Models\ContentType;
 
-require_once __DIR__."/assets.php";
 require_once __DIR__."/../../config.php";
 
 /*
